\name{Combo-class}
\docType{class}
\title{S4-classes for Accessing Combinatorial Iterators}
\alias{ComboApply-class}
\alias{ComboRes-class}
\alias{Constraints-class}
\alias{Partitions-class}
%
\alias{$,Combo,ANY,ANY-method}
\alias{$,ComboApply,ANY,ANY-method}
\alias{$,ComboRes,ANY,ANY-method}
\alias{$,Constraints,ANY,ANY-method}
\alias{$,Partitions,ANY,ANY-method}
%
\description{
  The \code{Combo} class family are S4-classes.
}

\section{Slots}{
  slots
}

\examples{
  showClass("Combo")
}
\keyword{classes}
\keyword{combinations}
\keyword{permutations}
\keyword{partitions}
\keyword{combinatorics}
