\name{comboGroups}
\alias{comboGroups}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Blah BLah
}
\description{
Blah Blah
}
\usage{
comboGroups(v, numGroups, retType = "3Darray", 
            lower = NULL, upper = NULL, Parallel = FALSE, nThreads = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{v}{
Blah Blah
}
  \item{numGroups}{
Blah Blah
}
  \item{retType}{
Blah Blah
}
  \item{lower}{
Blah Blah
}
  \item{upper}{
Blah Blah
}
  \item{Parallel}{
Blah Blah
}
  \item{nThreads}{
Blah Blah
}
}
\details{
Blah Blah
}
\value{
Blah Blah
}
\references{
Blah Blah
}
\author{
Blah Blah
}
\note{
Blah Blah
}
\examples{
comboGroups(8, 2)
}
